dialogues/requirements.txt
# Project dependencies
openai
pyautogen
python-dotenv
